conf/scripts: fix Onload headers processing
Make the script fail with an unhandled error, and deal with each
of them.
The 'scp_dir' function is not defined either in the sapi-ts repository
or in ts-conf, so a new 'rsync_from' function is introduced.
It is likely that the 'scp_dir' function is no longer in use.

OL-Redmine-Id: 12394
Signed-off-by: Damir Mansurov <damir.mansurov@oktetlabs.ru>
Acked-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>