scripts: move processing hosts to a right place
Some checks in the scripts/run.sh script rely on the fact that the TE
specific variables are contained in some file. Since the internal
structure of this repository is not regulated, this does not work in
all cases. These checks can be painlessly postponed until the necessary
variables are set.

OL-Redmine-Id: 12265
Signed-off-by: Damir Mansurov <damir.mansurov@oktetlabs.ru>
Acked-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>