scripts: fix epollN + safe options order issue
Replace "safe + epollN" options with a single "safe_epollN" option.
It fixes the issue related to the order of these options - if "epollN"
comes before "safe" we get a TRC tag that does not match the actual
behavior, because safe profile does not overwrite EF_UL_EPOLL variable,
but still adds the tag. That breaks epoll tests expectations.

OL-Redmine-Id: 12437
Signed-off-by: Sergey Nikitin <sergey.nikitin@oktetlabs.ru>
Reviewed-by: Alexandra Kossovsky <alexandra.kossovsky@oktetlabs.ru>