sockopts/ip_mtu_discover_udp: resend packet to update PMTU
During Path MTU discovery in test ip_mtu_discover_udp, Onload may need
more than one ICMP message to refresh Path MTU correctly. In that case,
the packet is resent until PMTU gets updated (up to 5 attempts are
made), and the verdict is printed.

In case of non-connected socket, we need to send a small packet from
aux_s (a connected socket) before getting the updated IP_MTU. A verdict
is printed in this case; it is added to TRC.

AMD-Jira-Id: ST-2518
Signed-off-by: Boris Shleyfman <bshleyfman@oktet.co.il>
Reviewed-by: Yurij Plotnikov <yurij.plotnikov@arknetworks.am>
Reviewed-by: Denis Pryazhennikov <denis.pryazhennikov@arknetworks.am>