sockapi-ts/lib: fix returned value in not pure netns check
Implementation of sockts_not_pure_netns_used() function didn't fit
its description and aims: the function returned TRUE in case of ipvlan
with NO netns but should return FALSE. We hadn't faced with this case
that could lead the function to misbehaviour due to ool_fix_consistency
added netns_iut in case of ipvlan (see ST-2119) thus the mistake had no
influence on sapi-ts behaviour and the proposing patch shouldn't do it
too. We have moved a parentheses inside returned boolean expression to
fit the intending behaviour, deleted outer parentheses to avoid the
ripples in the eyes and casted to bool the value returned by last
getenv in the expression.

Fixes: 0be5da288ae9 ("Sapi-TS going public")
Signed-off-by: Pavel Liulchak <pavel.liulchak@oktetlabs.ru>
Reviewed-by: Damir Mansurov <damir.mansurov@oktetlabs.ru>