ts: move function to common library place
Add function sockts_connect_retry() to lib/sockapi-ts_tcp, which is
common in several tests. It is used for making many attempts to
connect to IUT RPC server, in a loop.

Delete the function test_connect_loop() from basic/accept_multiple.
Replace it with sockts_connect_retry(), which is the rewritten
version of this function.

Signed-off-by: Boris Shleyfman <bshleyfman@oktet.co.il>
Reviewed-by: Yurij Plotnikov <yurij.plotnikov@arknetworks.am>