multicast/diff_vlan_mcast_check_id: implement new test
Check that VLAN does not receive wrong packets. Create two VLAN
interfaces on IUT. Join a multicasting group on the first VLAN with a
UDP socket, so this socket receives packets from this VLAN interface
only. Check that the socket only receives packets with correct VLAN
tag, equal to VLAN ID of the first VLAN interface.

AMD-Jira-Id: ST-2420
Signed-off-by: Boris Shleyfman <bshleyfman@oktet.co.il>
Reviewed-by: Yurij Plotnikov <yurij.plotnikov@arknetworks.am>
Reviewed-by: Denis Pryazhennikov <denis.pryazhennikov@arknetworks.am>