bnbvalue/func_socket_pipe_too_many: turn off close + caching
In too many sockets tests connections are being accepted until the
limit of opened file descriptors in the system is achieved (EMFILE
error). The last (not accepted) connection is saved in accept queue.
Tests close one of the accepted connections and accept the one saved
in the queue.

In case of socket caching, closing one of accepted connections does
not result in the freeing of accept queue and accept() still fails
with EMFILE error.

After Onload started to drop SYN in case of full accept queue, calling
connect() in this case started to result in timeout error. Tests with
many sockets and socket caching began to fail.

All in all, too many sockets tests does not make sense in case of
socket caching, because the accept queue is not freed. That is why this
combination should be turned off.

Link: https://github.com/Xilinx-CNS/onload/pull/182
OL-Redmine-Id: 13317
Signed-off-by: Alexander Kuzminov <alexander.kuzminov@oktetlabs.ru>
Reviewed-by: Alexander Dubovoj <alexander.dubovoj@oktetlabs.ru>
Reviewed-by: Damir Mansurov <damir.mansurov@oktetlabs.ru>
Reviewed-by: Alexandra Kossovsky <alexandra.kossovsky@oktetlabs.ru>